✅ split ChangePwView code from Setting screen
